Move wallpaper functionality into seperate file
